+++ AbnormalMouse/AbnormalMouse/Features/Scroll/MoveToScrollSettingsView.swift
@@ -80,6 +80,7 @@ private struct MoveToScrollSettingsView: View {
                     send: { .setNumberOfTapsRequired($0) }
                 ),
                 hasConflict: viewStore.hasConflict,
+                invalidReason: viewStore.invalidReason,
                 title: { Text(_L10n.View.activationKeyCombinationTitle) }
             )
         }
@@ -195,6 +196,7 @@ private struct HalfPageScrollView: View {
                         send: { .setNumberOfTapsRequired($0) }
                     ),
                     hasConflict: viewStore.hasConflict,
+                    invalidReason: viewStore.invalidReason,
                     title: { Text(_L10n.View.activationKeyCombinationTitle) }
                 )