+++ AbnormalMouse/AbnormalMouse/OverrideController/ZoomAndRotateController.swift
@@ -104,16 +104,17 @@ final class ZoomAndRotateController: OverrideController {
     }
 
     private func updateSettings() {
+        isActive = false
         state.zoomGestureDirection = persisted.zoomGestureDirection
         state.rotateGestureDirection = persisted.rotateGestureDirection
-        tapHold.keyCombination = persisted.keyCombination
+        tapHold.keyCombination = persisted.keyCombination?.validated
         tapHold.numberOfTapsRequired = persisted.numberOfTapsRequired
         if persisted.smartZoom.useZoomAndRotateDoubleTap {
             tap.numberOfTapsRequired = persisted.numberOfTapsRequired + 1
-            tap.keyCombination = persisted.keyCombination
+            tap.keyCombination = persisted.keyCombination?.validated
         } else {
             tap.numberOfTapsRequired = persisted.smartZoom.numberOfTapsRequired
-            tap.keyCombination = persisted.smartZoom.keyCombination
+            tap.keyCombination = persisted.smartZoom.keyCombination?.validated
         }
     }
 }
@@ -158,6 +159,7 @@ extension ZoomAndRotateController {
         case .mayBegin:
             state.zoomThreshold += abs(zoom)
             state.rotateThreshold += abs(rotate)
+            CGWarpMouseCursorPosition(state.mouseLocation)
         case let .begin(type):
             tapHold.consume()
             switch type {
@@ -167,6 +169,7 @@ extension ZoomAndRotateController {
                 p.postRotation(direction: rotateDirection, t: rotate, phase: .began)
             }
             p.postTranslation(phase: .began)
+            CGWarpMouseCursorPosition(state.mouseLocation)
         case let .hasBegun(type):
             switch type {
             case .zoom:
@@ -175,6 +178,7 @@ extension ZoomAndRotateController {
                 p.postRotation(direction: rotateDirection, t: rotate, phase: .changed)
             }
             p.postTranslation(phase: .changed)
+            CGWarpMouseCursorPosition(state.mouseLocation)
         case let .shouldEnd(type):
             switch type {
             case .zoom:
@@ -187,8 +191,6 @@ extension ZoomAndRotateController {
             state.zoomThreshold = 0
             state.rotateThreshold = 0
         }
-
-        CGWarpMouseCursorPosition(state.mouseLocation)
     }
 }